"/Java_sources/org/example/Configuration_Master/Configuration_Master_engine.java", "/Java_sources/org/example/Configuration_Master/Configuration_Master_server.java", & "/Java_sources/org/example/IPv4_client_authorization/IPv4_client_authorization_engine.java": made the verbosity-related integers be of Java type "long" rather than "short" [hah! ;-)], so that moderately-huge verbosity-level inputs such as the one in the command "./run.bash verbosity=100000" won`t cause the verbosity level to [silently, i.e. not crashing on a bad input :-(] wrap around and become _negative_ thus causing _all_ the verbosity comparisons I have written to yield false, i.e. when negative it is effectively equivalent to verbosity=0 [zero]

+++ Java_sources/org/example/Configuration_Master/Configuration_Master_engine.java
@@ -403,7 +403,7 @@ public class Configuration_Master_engine {
   // the next 3 lines: so non-ctor methods will be able to read these values without me needing to pass them around
   private boolean strict_checking_mode_enabled;
   private short   strictness_level;
-  private short   verbosity;
+  private long    verbosity;
 
 
   // start of ctor
@@ -411,7 +411,7 @@ public class Configuration_Master_engine {
                                debugFriendly_buffered_input   maturityLevel_aliases_input,
                                debugFriendly_buffered_input[] schema_inputs,
                                debugFriendly_buffered_input[] config_inputs,
-                               short                          verbosity_in,
+                               long                           verbosity_in,
                                short                          strictness_level___in,
                                boolean                        allow_empty_schema,
                                boolean                        allow_no_configurations
